feat(gym): per-gym modifier weight variants
Add modifier_variants JSONB field to Gym model so users can
declare available weights per weighted modifier (e.g. weighted
vest: 6/8/10kg, resistance bands: 15/25/35/45kg).

- Change weight_kg from int to float64 across the full stack
- Make resistance bands is_weighted with -0.6/kg impact
- Expose is_weighted flag via GET /equipment response
- Pass available weights to LLM prompt per modifier
- Add gym form UI for configuring weight variants
- Manual SQL migration for column type change

+++ app/lib/models/activity.dart
@@ -20,7 +20,7 @@ class Activity {
   @JsonKey(name: 'reps')
   final int reps;
   @JsonKey(name: 'weight_kg')
-  final int weightKg;
+  final double weightKg;
   @JsonKey(name: 'modifiers', defaultValue: [])
   final List<String> modifiers;
   @JsonKey(name: 'rest')
